optimizing network traffic - use [u8;4] instead of [f32;4] for colors - send only model instance index to browsers, and cache their matrices

+++ src/configs.rs
@@ -35,7 +35,7 @@ pub struct DevConfigStatsPlayer {
 
 #[derive(Debug, Deserialize)]
 pub struct DevConfigConsole {
-    pub color: [f32; 4],
+    pub color: [u8; 4],
 }
 
 #[derive(Debug, Deserialize)]
